Cloud Engineering/frontend/requirements.txt
pandas==2.2.2
numpy==1.26.4
PyYAML==6.0.1
boto3==1.35.99
matplotlib==3.9.2
seaborn==0.13.2
scikit-learn==1.5.1
scikit-surprise==1.1.4
SQLAlchemy==2.0.34
streamlit==1.45.1
python-dotenv==1.0.1
